app/backend/requirements.txt
torch
gradio
transformers
fastapi
uvicorn
pydantic
python-dotenv